+++ src/components/common/Header.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { S } from "./styled";
-import DropDown from "./DropDown/DropDown";
+import DropDown from "../DropDown/DropDown";
 import { useOncClickIcon } from "@/hooks/useOnClickIcon";
 import { useDropDownControl } from "@/hooks/useDropDownControl";
 const Header = () => {
